app/requirements/production.txt
celery==5.3.6
Django==5.0.3
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
loguru==0.7.2
psycopg2-binary==2.9.9
uWSGI==2.0.24
